useContext/createContext to store incomes and expenses

+++ src/components/ExpenseItem.tsx
@@ -1,24 +1,28 @@
-import React from 'react';
+import React, {useContext} from 'react';
 import {MoneyItem} from "../MoneyItem";
+import {GlobalContext} from "../context/GlobalState";
 
 interface ExpenseItemsProps {
     expense: MoneyItem;
     index: number;
-    deleteExpense: (i: any) => any
 }
 
-function ExpenseItem({expense, index, deleteExpense}: ExpenseItemsProps) {
+function ExpenseItem({expense, index}: ExpenseItemsProps) {
+    const {deleteExpense} = useContext(GlobalContext)
+
     let day = expense.date.getDate();
     let month = expense.date.getMonth() + 1;
     let year = expense.date.getFullYear();
 
-    const onRemoveClick = (i: number) => {
-        deleteExpense(i);
+    const onRemoveClick = (i: string) => {
+        if (deleteExpense) {
+            deleteExpense(i);
+        }
     }
 
     return (
         <div className="item">
-            <button className="remove-item" onClick={() => onRemoveClick(index)}>x</button>
+            <button className="remove-item" onClick={() => onRemoveClick(expense.id)}>x</button>
             <div className="desc">{expense.description}</div>
             <div className="price">${expense.price}</div>
             <div className="date">{day + "/" + month + "/" + year}</div>
